K.I.S.S. the tests, and on the way gain compatibility with WebStorm
