+++ Makefile
@@ -1,22 +1,17 @@
 
-TESTS = $(shell find test/ -name '*.test.js')
 DOCS_ = $(shell find lib/ -name '*.js')
 DOCS = $(DOCS_:.js=.json)
 DOCFILE = docs/source/_docs
 STABLE_BRANCH = 3.8.x
 
 test:
-	@MONGOOSE_DISABLE_STABILITY_WARNING=1 node test/dropdb.js
-	@MONGOOSE_DISABLE_STABILITY_WARNING=1 ./node_modules/.bin/mocha $(T) --async-only $(TESTS)
-	@MONGOOSE_DISABLE_STABILITY_WARNING=1 node test/dropdb.js
+	@MONGOOSE_DISABLE_STABILITY_WARNING=1 ./node_modules/.bin/mocha $(T) --async-only test/**/*.test.js
 
 test-short:
-	@MONGOOSE_DISABLE_STABILITY_WARNING=1 node test/dropdb.js
-	@MONGOOSE_DISABLE_STABILITY_WARNING=1 ./node_modules/.bin/mocha $(T) -g LONG -i --async-only $(TESTS)
-	@MONGOOSE_DISABLE_STABILITY_WARNING=1 node test/dropdb.js
+	@MONGOOSE_DISABLE_STABILITY_WARNING=1 ./node_modules/.bin/mocha $(T) -g LONG -i --async-only test/**/*.test.js
 
 test-long:
-	@MONGOOSE_DISABLE_STABILITY_WARNING=1 ./node_modules/.bin/mocha $(T) -g LONG --async-only $(TESTS)
+	@MONGOOSE_DISABLE_STABILITY_WARNING=1 ./node_modules/.bin/mocha $(T) -g LONG --async-only test/**/*.test.js
 
 docs: ghpages merge_stable docclean gendocs
 docs_all: docs_unstable docs
