+++ package.json
@@ -16,7 +16,7 @@
       , "mquery" : "0.7.0"
     }
   , "devDependencies": {
-        "mocha": "1.12.0"
+        "mocha": "1.20.0"
       , "node-static": "0.5.9"
       , "dox": "0.3.1"
       , "jade": "0.26.3"
@@ -29,7 +29,7 @@
       , "underscore": "1.5.2"
     }
   , "directories": { "lib": "./lib/mongoose" }
-  , "scripts": { "test": "make test" }
+  , "scripts": { "test": "set MONGOOSE_DISABLE_STABILITY_WARNING=1 && mocha --async-only test/**/*.test.js" }
   , "main": "./index.js"
   , "engines": { "node": ">=0.6.19" }
   , "bugs": {
