--- test/dropdb.js
@@ -1,20 +1,0 @@
-var start = require('./common')
-var db = start();
-db.once('open', function () {
-
-  // drop the default test database
-  db.db.dropDatabase(function () {
-    var db2 = db.useDb('mongoose-test-2');
-    db2.db.dropDatabase(function() {
-      // drop mongos test db if exists
-      var mongos = process.env.MONGOOSE_MULTI_MONGOS_TEST_URI;
-      if (!mongos) return process.exit();
-
-
-      var db = start({ uri: mongos, mongos: true });
-      db.once('open', function () {
-        db.db.dropDatabase(process.exit);
-      });
-    });
-  });
-});
